+++ src/apis/hooks/useWorkspaces.ts
@@ -1,8 +1,13 @@
 import { useQuery } from "@tanstack/react-query";
-import { getAllWorkspaceInvitations, getAllUserWorkspaces } from "@/apis/WorkspaceApi";
+import {
+  getAllWorkspaceInvitations,
+  getAllUserWorkspaces,
+  getSentWorkspaceInvitations,
+} from "@/apis/WorkspaceApi";
 
 const USER_WORKSPACES_QUERY_KEY = "user-workspaces" as const;
 const INVITATIONS_WORKSPACES_QUERY_KEY = "workspace-invitations" as const;
+const SENT_INVITATIONS_WORKSPACES_QUERY_KEY = "workspace-sent-invitations" as const;
 
 export const useWorkspaces = () =>
   useQuery({
@@ -15,5 +20,12 @@ export const useWorkspaceInvitations = () =>
   useQuery({
     queryKey: [INVITATIONS_WORKSPACES_QUERY_KEY],
     queryFn: () => getAllWorkspaceInvitations(),
+    staleTime: 5 * 60 * 1000,
+  });
+
+export const useSentWorkspaceInvitations = () =>
+  useQuery({
+    queryKey: [SENT_INVITATIONS_WORKSPACES_QUERY_KEY],
+    queryFn: () => getSentWorkspaceInvitations(),
     staleTime: 5 * 60 * 1000,
   });