+++ src/enums/InvitationStatusEnum.ts
@@ -2,6 +2,7 @@ export const InvitationStatusEnum = {
   PENDING: "PENDING",
   ACCEPTED: "ACCEPTED",
   REJECTED: "REJECTED",
+  CANCELLED: "CANCELLED",
 } as const;
 export type InvitationStatusEnum =
   (typeof InvitationStatusEnum)[keyof typeof InvitationStatusEnum];